Replace Environment with a HealthCheckRegistry
Before this change to create an instance of atam4j you had to pass in a dropwizard environment instance in order to register a health check and create a scheduler. I've replaced the need for an Environment with a HealthCheckRegistry instance and rolled our own scheduled executor. I've kept the dropwizard core dependency since we may need to expose http endpoints in the future to perform tasks/render views.